requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-login
requests
python-dotenv
gunicorn